Add single type formatting

+++ plugins/DotNetTools/perfpage.c
@@ -1076,7 +1076,7 @@ INT_PTR CALLBACK DotNetPerfPageDlgProc(
                                         // .NET 2/3/4 can sometimes show the TimeInJit value above 100% (dmex)
                                         // https://github.com/dotnet/coreclr/blob/ef1e2ab328087c61a6878c1e84f4fc5d710aebce/src/gc/gcee.cpp#L324
 
-                                        PhInitFormatF(&format[0], (context->DotNetPerfJit.timeInJit << 8) * 100 / (DOUBLE)(context->DotNetPerfJit.timeInJitBase << 8), 2);
+                                        PhInitFormatF(&format[0], (context->DotNetPerfJit.timeInJit << 8) * 100.f / (FLOAT)(context->DotNetPerfJit.timeInJitBase << 8), 2);
 
                                         if (PhFormatToBuffer(format, RTL_NUMBER_OF(format), formatBuffer, sizeof(formatBuffer), NULL))
                                         {
@@ -1502,7 +1502,7 @@ INT_PTR CALLBACK DotNetPerfPageDlgProc(
                                         PH_FORMAT format[1];
                                         WCHAR formatBuffer[10];
 
-                                        PhInitFormatF(&format[0], (DOUBLE)context->DotNetPerfGC.timeInGC * 100 / (DOUBLE)context->DotNetPerfGC.timeInGCBase, 2);
+                                        PhInitFormatF(&format[0], (FLOAT)context->DotNetPerfGC.timeInGC * 100.f / (FLOAT)context->DotNetPerfGC.timeInGCBase, 2);
 
                                         if (PhFormatToBuffer(format, RTL_NUMBER_OF(format), formatBuffer, sizeof(formatBuffer), NULL))
                                         {
@@ -1727,7 +1727,7 @@ INT_PTR CALLBACK DotNetPerfPageDlgProc(
                                         PH_FORMAT format[1];
                                         WCHAR formatBuffer[10];
 
-                                        PhInitFormatF(&format[0], (DOUBLE)context->DotNetPerfSecurity.timeRTchecks * 100 / (DOUBLE)context->DotNetPerfSecurity.timeRTchecksBase, 2);
+                                        PhInitFormatF(&format[0], (FLOAT)context->DotNetPerfSecurity.timeRTchecks * 100.f / (FLOAT)context->DotNetPerfSecurity.timeRTchecksBase, 2);
 
                                         if (PhFormatToBuffer(format, RTL_NUMBER_OF(format), formatBuffer, sizeof(formatBuffer), NULL))
                                         {
